base/views/product_views.py
from django.shortcuts import render

from rest_framework.decorators import api_view, permission_classes
from rest_framework.response import Response
from base.models import Product
from base.serializers import ProductSerializer
from rest_framework import status


@api_view(['GET'])
def getProducts(request):
    query = request.query_params.get('keyword')
    if query == None:
        query = ''
    products = Product.objects.filter(
        name__icontains=query).order_by('-createdAt')
    serializer = ProductSerializer(products, many=True)
    return Response({'products': serializer.data})
